[refactor] refactor file & code.
